Add expected snapshot for ambiguous path `receive`
This test should currently pass, but it, like the other journey
tests of the `receive` subcommand, runs only on CI. This is an
expected snapshot, so that if the test, code under test, or
expectation expressed in the snapshot is wrong, then it will be
revealed that there is a problem.

This also expresses the expectation to humans, which may help in
fixing the new test if it turns out to be wrong. Note however that
this does mean that the journey snapshot here is not strictly being
used for approval testing.