Add expected snapshots for ambiguous username journey tests
The tests appear to work, at least partially, in that a test
failure, of the kind expected due to gix-transport not yet checking
for usernames that ssh clients can interpret as an option argument,
is observed now that the current output has snapshots to compare to
(the comparison fails, which at this point is intended).

Not all tests ran, because of the runner script's fail-fast
behavior and because the `receive` subcommand tests only run on CI.

While I believe having these snapshots in now will make it eaiser
to ensure nothing essential was forgotten, once ambiguous username
handling is implemented, this does make these journey tests no
longer strictly approval tests (since the snapshots incorporate
foreknowledge of the intended behavior).