Start on demo script making repo with ../… filename
The repo this script makes contains a filename with a slash character
in it that, if not rejected, will create a file above the working tree.

This is a modification of make_traverse_dotgit_slashes.sh. Both
require some further revision, and since most of their content is
duplicated, it may be worthwhile to combine them to avoid that.