Use Dependabot version updates but just for GitHub Actions
A number of reusable actions, such as actions/checkout, are in old
major versions, including versions that use old deprecated versions
of Node.js, such as Node 16 (rather than the current Node 20).

Dependabot version updates were disabled for Rust crates in #143
and #144, with other means used to keep things up to date. This
does not reverse that even partially, as it enables version updates
only for GitHub Actions (which it has not been used on before).

Because most actions are specified with major version tags that get
reset to the latest stable minor/patch version under that major
version, the actual frequency of updates should not be anywhere
near daily. Nonetheless, it may make sense to further refine this
configuration, such as to check for updates less often as well as
to group multiple updates into a single PR when doing so.