Remove LFS CI step and modify docs/comments
This modifies the readme and one source code file to no longer say
that Git LFS is required to get generated archives, because they
are checked in as regular blobs instead these days, ever since
35439de (#1176). This may simplfy onboarding since readers are no
longer told they must do anything with `git lfs` to get started.

It also removes the LFS step from a CI test job definition that had
it, for the same reason: the generated archives are no longer
stored in LFS, so the only current expected effects of that step
are to make the job slightly more complicated and very slightly
slower. See also 7b3104d (#1360) and c60f270.