feat!: `gix-config` convenience initiative
A new `gix_config::AsKey` trait allows all `gix_config::File` methods
to receive an implementation of `AsKey` instead of the tuple of
`(section, subsection, value_name)`. This is the default, and all
methods that take a tuple have been renamed to `<name>_by()`, note the `_by` suffix.

The terminology was reworked so that `key` is now only used to identify a value.

This change also affects the public API of `gix`, which provides
ways to set values on configuration snapshots which now by default
will only take a `AsKey` implementation.

Note that `gix::config::tree::*` keys now also implement `AsKey`,
which allows them to be used more conveniently when setting values,
too.