Use `<<` rather than `<<-` heredoc operator
Because `<<-` only makes a difference when the contents of the here
document are indented with tabs, which was not being done anywhere
that `<<-` was used. See #1423 for details.

This also removes some spurious indentation (outside of the here
document) in make_dangling_symlink.sh and changes its blank lines
for readability and to match the style in similar scripts.